+++ src/preview.ts
@@ -1,25 +1,26 @@
 import { join } from 'path';
 import {
-  Disposable,
   Webview,
   WebviewPanel,
   window,
   Uri,
   ViewColumn,
-  ExtensionContext,
   QuickPickItemKind,
   DebugConfiguration,
   debug,
+  commands,
 } from 'vscode';
 import { exState } from './wn-tree-provider';
 import { getSetting, setSetting, WorkspaceSetting } from './workspace-state';
 import { debugSkipFiles, openUri } from './utilities';
 import { cancelLastOperation } from './tasks';
+import { nexusURL } from './webview-debug';
 
 enum MessageType {
   setMobile = 'setMobile',
   setWeb = 'setWeb',
   device = 'device',
+  qr = 'qr',
   stopSpinner = 'stopSpinner',
 }
 
@@ -69,10 +70,17 @@ export function viewInEditor(
         enableScripts: true,
         retainContextWhenHidden: true,
       });
+
+  // Allows QR Codes to be displayed
+  const onDiskPath = Uri.file(join(exState.context.extensionPath, 'resources', 'qrious.min.js'));
+  const qrSrc = panel.webview.asWebviewUri(onDiskPath);
   lastUrl = url;
   const extensionUri = exState.context.extensionUri;
-  panel.webview.html = url ? getWebviewContent(panel.webview, extensionUri) : '';
+  panel.webview.html = url ? getWebviewContent(panel.webview, extensionUri, qrSrc.toString()) : '';
   panel.iconPath = iconFor('globe');
+  if (!existingPanel) {
+    commands.executeCommand('workbench.action.closeSidebar');
+  }
   let device = getSetting(WorkspaceSetting.emulator);
 
   if (overrideAsWeb) {
@@ -99,6 +107,15 @@ export function viewInEditor(
         panel.webview.postMessage({ command: MessageType.device, device });
         return;
       }
+    }
+    if (message.command == 'browser') {
+      openUri(lastUrl);
+      return;
+    }
+    if (message.command == 'qr') {
+      const item = nexusURL(lastUrl);
+      panel.webview.postMessage({ command: MessageType.qr, item });
+      return;
     }
     if (message.command == 'add') {
       console.log('add');
@@ -188,79 +205,7 @@ async function selectMockDevice(): Promise<device> {
   return devices.find((device) => selected.includes(device.name));
 }
 
-// export class PreviewPanel {
-//   public static currentPanel: PreviewPanel | undefined;
-//   private readonly panel: WebviewPanel;
-//   private disposables: Disposable[] = [];
-
-//   private path: string;
-
-//   private constructor(panel: WebviewPanel, extensionUri: Uri, path: string, context: ExtensionContext) {
-//     if (!path) {
-//       path = extensionUri.fsPath;
-//     }
-//     this.panel = panel;
-//     this.path = path;
-//     this.panel.onDidDispose(() => this.dispose(), null, this.disposables);
-//     this.panel.webview.html = this.getWebviewContent(this.panel.webview, extensionUri);
-//     this.setWebviewMessageListener(this.panel.webview, extensionUri, path, context);
-//   }
-
-//   public static init(extensionUri: Uri, path: string, context: ExtensionContext, force?: boolean) {
-//     if (PreviewPanel.currentPanel) {
-//       // If the webview panel already exists reveal it
-//       PreviewPanel.currentPanel.panel.reveal(ViewColumn.One);
-//     } else {
-//       // If a webview panel does not already exist create and show a new one
-//       const panel = window.createWebviewPanel(
-//         // Panel view type
-//         'preview',
-//         // Panel title
-//         'Preview',
-//         ViewColumn.One,
-//         {
-//           enableScripts: true,
-//           localResourceRoots: [Uri.joinPath(extensionUri, 'out'), Uri.joinPath(extensionUri, 'preview', 'build')],
-//         },
-//       );
-
-//       PreviewPanel.currentPanel = new PreviewPanel(panel, extensionUri, path, context);
-//     }
-//   }
-
-//   public dispose() {
-//     PreviewPanel.currentPanel = undefined;
-//     this.panel.dispose();
-//     while (this.disposables.length) {
-//       const disposable = this.disposables.pop();
-//       if (disposable) {
-//         disposable.dispose();
-//       }
-//     }
-//   }
-
-//   private setWebviewMessageListener(webview: Webview, extensionUri: Uri, path: string, context: ExtensionContext) {
-//     webview.onDidReceiveMessage(
-//       async (message: any) => {
-//         const command = message.command;
-//         switch (command) {
-//           case MessageType.setMobile: {
-//             webview.postMessage({ command });
-//             break;
-//           }
-//           case MessageType.setWeb: {
-//             webview.postMessage({ command });
-//             break;
-//           }
-//         }
-//       },
-//       undefined,
-//       this.disposables,
-//     );
-//   }
-// }
-
-function getWebviewContent(webview: Webview, extensionUri: Uri) {
+function getWebviewContent(webview: Webview, extensionUri: Uri, qrSrc: string) {
   const stylesUri = getUri(webview, extensionUri, ['preview', 'build', 'styles.css']);
   const runtimeUri = getUri(webview, extensionUri, ['preview', 'build', 'runtime.js']);
   const polyfillsUri = getUri(webview, extensionUri, ['preview', 'build', 'polyfills.js']);
@@ -279,6 +224,7 @@ function getWebviewContent(webview: Webview, extensionUri: Uri) {
         <link rel="stylesheet" type="text/css" href="${stylesUri}">
         <titlePreview</title>
       </head>
+      <script src="${qrSrc}"></script>
       <body>
         <app-root></app-root>
         <script type="module" nonce="${nonce}" src="${runtimeUri}"></script>